+++ Revitalize/models.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from json import JSONDecodeError
 
 from django.contrib.auth.models import AbstractUser
 from django.core.validators import RegexValidator
@@ -13,7 +12,10 @@ from Revitalize.data_analysis_system import DataAnalysisSystem
 
 import logging
 
-from Revitalize.typing import Numeric, DataPoint
+from Revitalize.language_support import database_to_string
+from Revitalize.typing import Numeric
+from Revitalize.utils import ModelHelper
+from Revitalize.validation import validate_json
 
 logger = logging.getLogger(__name__)
 _context = 'models.'
@@ -58,119 +60,6 @@ class User(AbstractUser):
                 points = q
             points = points.union(q)
         return points
-
-
-def validate_json(j: str):
-    try:
-        return json.loads(j) is not None
-    except JSONDecodeError as e:
-        raise ValidationError(e)
-
-
-def pre_validate_json(j: str):
-    try:
-        return json.loads(j)
-    except JSONDecodeError as e:
-        raise ValidationError(e)
-
-
-class LangCode(models.TextChoices):
-    UNKNOWN = '?', _('Unknown')
-    ENGLISH = 'EN', _('English')
-    FRENCH = 'FR', _('French')
-
-
-def _str(entry, lang=LangCode.ENGLISH.value, default=None):
-    s = entry.value if entry is not None else default
-    if lang == LangCode.ENGLISH.value:
-        # English is default
-        pass
-    elif lang == LangCode.FRENCH.value:
-        # TODO
-        pass
-    return s
-
-
-# This class is just a helper for dealing with some Django features quickly during development
-# and will at some point be removed. Please ignore.
-class ModelHelper:
-    model_fields: dict = {}
-    model_sorts: dict = {}
-    model_displays: dict = {}
-    model_filter_ons: dict = {}
-    model_search_ons: dict = {}
-    model_do_not_serializes: dict = {}
-    model_text_types: dict = {}
-    model_foreigns: dict = {}
-
-    admin_to_register: list = []
-
-    dicts = [model_fields, model_sorts, model_displays, model_filter_ons, model_search_ons, model_do_not_serializes,
-             model_text_types, model_foreigns]
-
-    @classmethod
-    def inherit(cls, parent, child):
-        for d in cls.dicts:
-            if parent in d.keys():
-                d[child] = d[parent].copy()
-            else:
-                if d is cls.model_fields:
-                    d[child] = []
-                else:
-                    d[child] = {}
-
-    @classmethod
-    def get_(cls, dic: dict, key):
-        if key not in dic.keys():
-            if dic is cls.model_fields:
-                dic[key] = []
-            else:
-                dic[key] = {}
-        return dic[key]
-
-    @classmethod
-    def register(cls, model, field, priority, to_display=True, to_filter=False, to_search=False, to_serialize=True,
-                 text_type=None, foreign=None):
-        model_fields: list = cls.get_(cls.model_fields, model)
-        model_sorts: dict = cls.get_(cls.model_sorts, model)
-        model_displays: dict = cls.get_(cls.model_displays, model)
-        model_filter_ons: dict = cls.get_(cls.model_filter_ons, model)
-        model_search_ons: dict = cls.get_(cls.model_search_ons, model)
-        model_do_not_serializes: dict = cls.get_(cls.model_do_not_serializes, model)
-        model_text_types: dict = cls.get_(cls.model_text_types, model)
-        model_foreigns: dict = cls.get_(cls.model_foreigns, model)
-
-        model_fields.append(field)
-        model_sorts[field] = priority
-        model_displays[field] = to_display
-        model_filter_ons[field] = to_filter
-        model_search_ons[field] = to_search
-        model_do_not_serializes[field] = not to_serialize
-        model_text_types[field] = text_type
-        model_foreigns[field] = foreign
-
-    @classmethod
-    def to_admin(cls, model):
-        cls.admin_to_register.append(model)
-
-    @classmethod
-    def serialize(cls, model):
-        if not isinstance(model, str):
-            key = model.__name__
-        else:
-            key = model
-
-        model_fields: list = cls.get_(cls.model_fields, key)
-        model_sorts: dict = cls.get_(cls.model_sorts, key)
-        model_do_not_serializes: dict = cls.get_(cls.model_do_not_serializes, key)
-        model_text_types: dict = cls.get_(cls.model_text_types, key)
-        model_foreigns: dict = cls.get_(cls.model_foreigns, key)
-
-        temp = [f for f in model_fields if not model_do_not_serializes[f]]
-
-        temp.sort(key=lambda x: model_sorts[x], reverse=True)
-
-        return temp
 
 
 class String(models.Model):
@@ -2919,8 +2808,8 @@ class Indicator(Analysable):
 
         if not self.dynamic:
             info['id'] = self.id
-            info['name'] = _str(self.name)
-            info['description'] = _str(self.description)
+            info['name'] = database_to_string(self.name)
+            info['description'] = database_to_string(self.description)
             info['max'] = self.max
             info['min'] = self.min
             info['target'] = self.target
